requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.6.1
httpx==0.25.2
python-multipart==0.0.6
beautifulsoup4==4.12.2
lxml==4.9.4
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
selenium==4.15.2
webdriver-manager==4.0.1
